tess_ocr/requirements.txt
asgiref==3.4.0
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
djongo==1.3.6
gunicorn==20.1.0
numpy==1.21.0
opencv-python==4.5.2.54
Pillow==8.2.0
pymongo==3.11.4
pytesseract==0.3.8
pytz==2021.1
six==1.16.0
sqlparse==0.2.4
whitenoise==5.2.0
